# docs/swagger.yaml
definitions:
  models.AvatarResponse:
    description: Response model for avatar upload
    properties:
      profile_image:
        example: https://example.com/avatar.jpg
        type: string
    type: object
...
  models.CreateCommentRequest:
    description: Request model for creating a new comment on a post
    properties:
      content:
        example: This is a great post!
        type: string
    required:
    - content
    type: object
...
  models.StandardResponse:
    description: A standard API response format with data payload
    properties:
      data: {}
      error:
        example: Invalid input
        type: string
      message:
        example: Operation completed successfully
        type: string
      status:
        example: success
        type: string
    type: object
...
  models.UpdateCommentRequest:
    description: Request model for updating an existing comment
    properties:
      content:
        example: This is my updated comment
        type: string
    required:
    - content
    type: object
...
info:
  contact: {}
paths:
...
  /comments/{commentID}:
    delete:
      description: Removes a comment from a post
      parameters:
      - description: Comment ID
        in: path
        name: commentID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: Comment not found
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      security:
      - BearerAuth: []
      summary: Delete a comment
      tags:
      - Comments
    put:
      consumes:
      - application/json
      description: Updates an existing comment
      parameters:
      - description: Comment ID
        in: path
        name: commentID
        required: true
        type: integer
      - description: Updated comment content
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated comment
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: Comment not found
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      security:
      - BearerAuth: []
      summary: Update a comment
      tags:
      - Comments
...
  /posts/{id}/comments:
    post:
      consumes:
      - application/json
      description: Adds a new comment to a post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment content
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/models.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created comment
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: Post not found
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      security:
      - BearerAuth: []
      summary: Create a new comment
      tags:
      - Comments
...
  /posts/{postID}/comments:
    get:
      description: Returns all comments for a specific post
      parameters:
      - description: Post ID
        in: path
        name: postID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of comments
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Get comments for a post
      tags:
      - Comments
...
  /profile/avatar:
    post:
      consumes:
      - multipart/form-data
      description: Upload a new avatar image for the current user
      parameters:
      - description: Avatar image file (JPG, JPEG, PNG, max 2MB)
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Avatar uploaded successfully
          schema:
            $ref: '#/definitions/models.AvatarResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      security:
      - BearerAuth: []
      summary: Upload user avatar
      tags:
      - Users
...
swagger: "2.0"